Move core.config.config to core.config
