+++ src/scout_apm/flask/__init__.py
@@ -2,7 +2,7 @@ from flask import current_app
 from flask.globals import _request_ctx_stack
 
 import scout_apm.core
-from scout_apm.core.config.config import ScoutConfig
+from scout_apm.core.config import ScoutConfig
 from scout_apm.core.tracked_request import TrackedRequest
 from scout_apm.core.monkey import CallableProxy
