+++ src/scout_apm/pyramid/__init__.py
@@ -1,6 +1,6 @@
 import scout_apm.core
 from scout_apm.core.tracked_request import TrackedRequest
-from scout_apm.core.config.config import ScoutConfig
+from scout_apm.core.config import ScoutConfig
 from scout_apm.api.context import Context
